first steps towards separation of concerns for matrix and tile copy
